changes interface for badge instance calls
a single function is exposed now, and depeding on the parameters passed, that
function will take care of filtering by delegating to other internal functions.
Not all the filters are available yet, but this is a way to relieve the client
of the burden of filtering locally, and also to pass back a lot less information
through the wire, which is always a good idea.